requirements.txt
discord.py
discord-ext-ipc
quart